Moved unit tests for feels like temperature auxiliary functions into single module and deleted old test files.
