--- improver_tests/feels_like_temperature/test_calculate_apparent_temperature.py
@@ -1,59 +1,0 @@
-# -*- coding: utf-8 -*-
-# -----------------------------------------------------------------------------
-# (C) British Crown Copyright 2017-2019 Met Office.
-# All rights reserved.
-#
-# Redistribution and use in source and binary forms, with or without
-# modification, are permitted provided that the following conditions are met:
-#
-# * Redistributions of source code must retain the above copyright notice, this
-#   list of conditions and the following disclaimer.
-#
-# * Redistributions in binary form must reproduce the above copyright notice,
-#   this list of conditions and the following disclaimer in the documentation
-#   and/or other materials provided with the distribution.
-#
-# * Neither the name of the copyright holder nor the names of its
-#   contributors may be used to endorse or promote products derived from
-#   this software without specific prior written permission.
-#
-# THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
-# AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
-# IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
-# ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE
-# LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
-# CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
-# SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
-# INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
-# CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
-# ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
-# POSSIBILITY OF SUCH DAMAGE.
-"""Unit tests for the feels_like_temperature.ApparentTemperature plugin."""
-
-import unittest
-
-import numpy as np
-from iris.tests import IrisTest
-
-from improver.feels_like_temperature import calculate_apparent_temperature
-
-from ..set_up_test_cubes import set_up_variable_cube
-
-
-class Test_calculate_apparent_temperature(IrisTest):
-    """Test the apparent temperature function."""
-
-    def test_values(self):
-        """Test output values from apparent temperature equation."""
-        temperature = np.full((1, 3), 22.)
-        wind_speed = np.full((1, 3), 5)
-        pressure = np.array([[99998., 101248., 102498.]])
-        relh = np.array([[0., 0.075, 0.15]])
-        expected_result = np.array([[16.9300, 17.3283, 17.7267]])
-        result = calculate_apparent_temperature(
-            temperature, wind_speed, relh, pressure)
-        self.assertArrayAlmostEqual(result, expected_result, decimal=4)
-
-
-if __name__ == '__main__':
-    unittest.main()
