--- improver_tests/feels_like_temperature/test_calculate_wind_chill.py
@@ -1,56 +1,0 @@
-# -*- coding: utf-8 -*-
-# -----------------------------------------------------------------------------
-# (C) British Crown Copyright 2017-2019 Met Office.
-# All rights reserved.
-#
-# Redistribution and use in source and binary forms, with or without
-# modification, are permitted provided that the following conditions are met:
-#
-# * Redistributions of source code must retain the above copyright notice, this
-#   list of conditions and the following disclaimer.
-#
-# * Redistributions in binary form must reproduce the above copyright notice,
-#   this list of conditions and the following disclaimer in the documentation
-#   and/or other materials provided with the distribution.
-#
-# * Neither the name of the copyright holder nor the names of its
-#   contributors may be used to endorse or promote products derived from
-#   this software without specific prior written permission.
-#
-# THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
-# AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
-# IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
-# ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE
-# LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
-# CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
-# SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
-# INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
-# CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
-# ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
-# POSSIBILITY OF SUCH DAMAGE.
-"""Unit tests for the feels_like_temperature.WindChill plugin."""
-
-import unittest
-
-import numpy as np
-from iris.tests import IrisTest
-
-from improver.feels_like_temperature import calculate_wind_chill
-
-from ..set_up_test_cubes import set_up_variable_cube
-
-
-class Test_calculate_wind_chill(IrisTest):
-    """Test the wind chill function."""
-
-    def test_values(self):
-        """Test output values when from the wind chill equation."""
-        temperature = np.full((1, 3), 1.7)
-        wind_speed = np.full((1, 3), 3) * 60 * 60 / 1000.
-        expected_result = np.full((1, 3), -1.4754, dtype=np.float32)
-        result = calculate_wind_chill(temperature, wind_speed)
-        self.assertArrayAlmostEqual(result, expected_result, decimal=4)
-
-
-if __name__ == '__main__':
-    unittest.main()
